Add Joplin installation script for HX99G
This commit adds a script (`install_joplin.txt`) to the `HX99G_Ubuntu_24.04.1_LTS_x86_64` directory for installing and configuring Joplin on the HX99G machine running Ubuntu 24.04.1 LTS x86_64.

The script automates the following:

- Installs Joplin using the official installation script.
- Provides instructions for configuring Joplin to synchronize with Dropbox.
- Guides users through enabling encryption and setting a master password.

This addition simplifies the process of setting up and using Joplin for note-taking and to-do list management on the specified machine configuration.